Introducing macros for log levels
I have added a macro for every loglevel the logger currently supports.
This allows us to provide information like calling file, func and line
number. The information is stored in a LogMessage object. I will have to
implement a parsing engine for conversion patterns to make use of this
new things.